update code for feature edit
