+++ src/assets/lang/en.json
@@ -26,6 +26,7 @@
     "loading": "Loading...",
     "refresh": "Refresh",
     "events": "Events",
+    "option": "Choose field to change",
     "password": "Password",
     "under_contruct": "Under Contruct"
   },
@@ -54,7 +55,7 @@
     "duplicate_param_name": "Duplicate param name.",
     "missing_required_params": "Missing required params !",
     "update_request_error": "Update request error !",
-    "wrong_password": "Invalid password!",
+    "wrong_password": "Invalid password!"
   },
   "menu": {
     "projects": "Projects",
@@ -100,7 +101,10 @@
     "param_key": "Param key",
     "add_param": "Add Param",
     "header_key": "Header key",
-    "add_header": "Add header"
+    "add_header": "Add header",
+    "accounts": "Accounts",
+    "contracts": "Contracts",
+    "general": "General"
   },
   "playground": {
     "new_process": "New Process",
