+++ src/modules/edit/containers/EditGeneral.js
@@ -1,7 +1,9 @@
+// eslint-disable-next-line
 import { connect } from 'react-redux'
 import EditGeneral from '../components/EditGeneral'
 import { descrypt, encrypt } from '../../../common/utils/encrypt'
 import { store } from '../../../common/utils/database'
+import saveAs from 'file-saver'
 import { addProject, removeProject } from '../../project/actions'
 import { loading } from '../../../common/middlewares/effects'
 import { setCurrentProject } from '../../dashboard/actions'
@@ -38,6 +40,11 @@ const mapDispatchToProps = (dispatch, props) => ({
         }))
         await store.removeItem(`project_${id}`)
         dispatch(removeProject(+id))
+        const blob = new Blob([JSON.stringify({
+          ...newDefaultData,
+          encrypted
+        })], { type: 'text/plain;charset=utf-8' })
+        saveAs(blob, `${newDefaultData.name}.json`)
         return true
       })
       return result
